Split HardwareUtility into CompressorSystem and PDH implementation
