+++ src/main/java/frc/robot/Robot.java
@@ -7,6 +7,8 @@ import com.revrobotics.CANSparkMax;
 import com.revrobotics.REVPhysicsSim;
 
 import edu.wpi.first.math.system.plant.DCMotor;
+import edu.wpi.first.wpilibj.Compressor;
+import edu.wpi.first.wpilibj.PneumaticsModuleType;
 // WPILib Imports
 import edu.wpi.first.wpilibj.TimedRobot;
 import edu.wpi.first.cscore.CvSink;
@@ -20,8 +22,8 @@ import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
 // Systems
 import frc.robot.systems.DriveFSMSystem;
 import frc.robot.systems.GrabberFSM;
-import frc.robot.wrappers.HardwareUtility;
 import frc.robot.systems.BallHandlingFSM;
+import frc.robot.systems.CompressorSystem;
 
 /**
  * The VM is configured to automatically run this class, and to call the functions corresponding to
@@ -29,7 +31,7 @@ import frc.robot.systems.BallHandlingFSM;
  */
 public class Robot extends TimedRobot {
 	private TeleopInput input;
-	private HardwareUtility hwUtilityController;
+	private Compressor pneumaticsCompressor;
 
 	// Systems
 	private DriveFSMSystem driveFsmSystem;
@@ -41,7 +43,7 @@ public class Robot extends TimedRobot {
 	private final int cameraBrightness = 25;
 	private final int camWidth = 320;
 	private final int camHeight = 240;
-	public static final boolean RUN_COMPRESSOR = false;
+	private static final boolean RUN_COMPRESSOR = false;
 
 	/**
 	 * This function is run when the robot is first started up and should be used for any
@@ -51,8 +53,8 @@ public class Robot extends TimedRobot {
 	public void robotInit() {
 		System.out.println("robotInit");
 
-		//Init hardware utility controller
-		hwUtilityController = new HardwareUtility(RUN_COMPRESSOR);
+		//Init Compressor
+		CompressorSystem compressorSystem = new CompressorSystem(RUN_COMPRESSOR);
 
 		//Init Driver Inputs
 		input = new TeleopInput();
@@ -155,14 +157,5 @@ public class Robot extends TimedRobot {
 	 */
 	public void updateShuffleboardVisualizations() {
 		SmartDashboard.updateValues();
-	}
-
-	/**
-	 * Get the hardware utility controller object to control
-	 * and read information from the PDH and Compressor.
-	 * @return HardwareUtility object
-	 */
-	public HardwareUtility getHardwareUtilityController() {
-		return hwUtilityController;
 	}
 }
